backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
insightface
onnxruntime
opencv-python
python-jose
passlib
python-dotenv
pydantic
pillow
psycopg2-binary
requests